# packages/service-backend/swagger.json
{"openapi":"3.0.0","paths":{"/health":{"get":{"operationId":"AppController_healthCheck","parameters":[],"responses":{"200":{"description":""}}}},"/alert":{"post":{"operationId":"AlertController_sendAlertToAll","parameters":[],"responses":{"200":{"description":""}}}},"/user/register":{"post":{"operationId":"UserController_registerUser","parameters":[],"requestBody":{"required":true,"description":"user data model","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserModel"},"examples":{"template":{"value":{"firstname":"name","lastname":"lastname","email":"test@test.test","password":"1234aaBB@"}}}}}},"responses":{"200":{"description":"user's token","content":{"application/json":{"schema":{"type":"string"}}}},"400":{"description":"Parameters are not valid"},"403":{"description":"User does not have permission to execute this action"}},"tags":["User"],"security":[{"cookie":[]}]}},"/user/login":{"post":{"operationId":"UserController_loginUser","parameters":[],"requestBody":{"required":true,"description":"user data model","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginUserModel"},"examples":{"template":{"value":{"email":"test@test.test","password":"1234aaBB@"}}}}}},"responses":{"200":{"description":"true","content":{"application/json":{"schema":{"type":"string"}}}},"400":{"description":"Parameters are not valid"},"403":{"description":"User does not have permission to execute this action"}},"tags":["User"],"security":[{"cookie":[]}]}},"/user":{"get":{"operationId":"UserController_getUser","parameters":[],"responses":{"200":{"description":"user's data","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetUserModel"}}}},"400":{"description":"Parameters are not valid"},"403":{"description":"User does not have permission to execute this action"}},"tags":["User"]},"put":{"operationId":"UserController_updateUser","parameters":[],"requestBody":{"required":true,"description":"user data model","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserModel"},"examples":{"template":{"value":{"firstname":"name","lastname":"lastname","email":"test@test.test","password":"1234"}}}}}},"responses":{"200":{"description":"user's token","content":{"application/json":{"schema":{"type":"string"}}}},"400":{"description":"Parameters are not valid"},"403":{"description":"User does not have permission to execute this action"}},"tags":["User"],"security":[{"cookie":[]}]},"delete":{"operationId":"UserController_deleteUser","parameters":[],"responses":{"200":{"description":"Ok.","content":{"application/json":{"schema":{"type":"string"}}}},"400":{"description":"Parameters are not valid"},"403":{"description":"User does not have permission to execute this action"}},"tags":["User"]}},"/user/courses":{"get":{"operationId":"UserController_getUserCourses","parameters":[],"responses":{"200":{"description":"list the courses of a user","content":{"application/json":{"schema":{"type":"string"}}}},"400":{"description":"Parameters are not valid"},"403":{"description":"User does not have permission to execute this action"}},"tags":["User"]}},"/course":{"post":{"operationId":"CourseController_postCourse","parameters":[],"requestBody":{"required":true,"description":"user data model","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCourseModel"},"examples":{"template":{"value":{"title":"Course Title","description":"Course description","picture":"Picture Data or Url"}}}}}},"responses":{"200":{"description":"course create response","content":{"application/json":{"schema":{"type":"string"}}}},"400":{"description":"Parameters are not valid"}},"tags":["Course"]},"delete":{"operationId":"CourseController_deleteCourse","parameters":[],"requestBody":{"required":true,"description":"user data model","content":{"application/json":{"schema":{"$ref":"#/components/schemas/IdCourseModel"},"examples":{"template":{"value":{"id":"id"}}}}}},"responses":{"200":{"description":"course delete response","content":{"application/json":{"schema":{"type":"string"}}}},"400":{"description":"Parameters are not valid"}},"tags":["Course"]}},"/course/{id}":{"get":{"operationId":"CourseController_getCourse","parameters":[{"name":"id","required":true,"in":"path","description":"Id of the course","schema":{"type":"string"}}],"responses":{"200":{"description":"course content response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CourseModel"}}}},"400":{"description":"Parameters are not valid"}},"tags":["Course"]},"put":{"operationId":"CourseController_updateCourse","parameters":[{"name":"id","required":true,"in":"path","description":"Id of the course","schema":{"type":"string"}}],"requestBody":{"required":true,"description":"user data model","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateCourseModel"},"examples":{"template":{"value":{"ownerId":"id","title":"Course Title","description":"Course description","picture":"Picture Data or Url"}}}}}},"responses":{"200":{"description":"course update response","content":{"application/json":{"schema":{"type":"string"}}}},"400":{"description":"Parameters are not valid"}},"tags":["Course"]}},"/course/sections/{id}":{"get":{"operationId":"CourseController_getCourseSections","parameters":[{"name":"id","required":true,"in":"path","description":"Id of the course","schema":{"type":"string"}}],"responses":{"200":{"description":"course's sections","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/SectionModel"}}}}},"400":{"description":"Parameters are not valid"}},"tags":["Course"]}},"/course/user/{id}":{"post":{"operationId":"CourseController_addUserToCourse","parameters":[{"name":"id","required":true,"in":"path","description":"Id of the course","schema":{"type":"string"}}],"responses":{"200":{"description":"user added to course response","content":{"application/json":{"schema":{"type":"string"}}}},"400":{"description":"Parameters are not valid"}},"tags":["Course"]}},"/section":{"post":{"operationId":"SectionController_registerSection","parameters":[],"requestBody":{"required":true,"description":"user data model","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateSectionModel"},"examples":{"template":{"value":{"courseId":"Course Id","title":"Section Title","description":"Section decsription"}}}}}},"responses":{"200":{"description":"section create response","content":{"application/json":{"schema":{"type":"string"}}}},"400":{"description":"Parameters are not valid"}},"tags":["Section"]},"delete":{"operationId":"SectionController_deleteSection","parameters":[],"requestBody":{"required":true,"description":"user data model","content":{"application/json":{"schema":{"$ref":"#/components/schemas/IdSectionModel"},"examples":{"template":{"value":{"id":"id"}}}}}},"responses":{"200":{"description":"section delete response","content":{"application/json":{"schema":{"type":"string"}}}},"400":{"description":"Parameters are not valid"}},"tags":["Section"]}},"/section/{id}":{"get":{"operationId":"SectionController_getSection","parameters":[{"name":"id","required":true,"in":"path","description":"Id of the section","schema":{"type":"string"}}],"responses":{"200":{"description":"section content response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SectionModel"}}}},"400":{"description":"Parameters are not valid"}},"tags":["Section"]},"put":{"operationId":"SectionController_updateSection","parameters":[{"name":"id","required":true,"in":"path","description":"Id of the section","schema":{"type":"string"}}],"requestBody":{"required":true,"description":"user data model","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateSectionModel"},"examples":{"template":{"value":{"courseId":"id","title":"Section Title","description":"Section decsription"}}}}}},"responses":{"200":{"description":"section update response","content":{"application/json":{"schema":{"type":"string"}}}},"400":{"description":"Parameters are not valid"}},"tags":["Section"]}},"/section/lessons/{id}":{"get":{"operationId":"SectionController_getSectionLessons","parameters":[{"name":"id","required":true,"in":"path","description":"Id of the section","schema":{"type":"string"}}],"responses":{"200":{"description":"section's lessons","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/LessonModel"}}}}},"400":{"description":"Parameters are not valid"}},"tags":["Section"]}},"/lesson":{"post":{"operationId":"LessonController_registerLesson","parameters":[],"requestBody":{"required":true,"description":"user data model","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateLessonModel"},"examples":{"template":{"value":{"section_id":"Section Id","title":"Lesson Title","description":"Lesson decsription"}}}}}},"responses":{"200":{"description":"lesson create response","content":{"application/json":{"schema":{"type":"string"}}}},"400":{"description":"Parameters are not valid"}},"tags":["Lesson"]},"delete":{"operationId":"LessonController_deleteLesson","parameters":[],"requestBody":{"required":true,"description":"user data model","content":{"application/json":{"schema":{"$ref":"#/components/schemas/IdLessonModel"},"examples":{"template":{"value":{"id":"id"}}}}}},"responses":{"200":{"description":"lesson delete response","content":{"application/json":{"schema":{"type":"string"}}}},"400":{"description":"Parameters are not valid"}},"tags":["Lesson"]}},"/lesson/{id}":{"get":{"operationId":"LessonController_getLesson","parameters":[{"name":"id","required":true,"in":"path","description":"Id of the lesson","schema":{"type":"string"}}],"responses":{"200":{"description":"lesson content response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LessonModel"}}}},"400":{"description":"Parameters are not valid"}},"tags":["Lesson"]},"put":{"operationId":"LessonController_updateLesson","parameters":[{"name":"id","required":true,"in":"path","description":"Id of the lesson","schema":{"type":"string"}}],"requestBody":{"required":true,"description":"user data model","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LessonModel"},"examples":{"template":{"value":{"sectionId":"id","title":"Lesson Title","description":"Lesson decsription"}}}}}},"responses":{"200":{"description":"lesson update response","content":{"application/json":{"schema":{"type":"string"}}}},"400":{"description":"Parameters are not valid"}},"tags":["Lesson"]}},"/lesson/questions/{id}":{"get":{"operationId":"LessonController_getLessonQuestions","parameters":[{"name":"id","required":true,"in":"path","description":"Id of the lesson","schema":{"type":"string"}}],"responses":{"200":{"description":"lesson's questions","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/QuestionModel"}}}}},"400":{"description":"Parameters are not valid"}},"tags":["Lesson"]}},"/lesson/{id}/join":{"post":{"operationId":"LessonController_joinLesson","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"description":"user data model","content":{"application/json":{"schema":{"$ref":"#/components/schemas/JoinLessonModel"},"examples":{"template":{"value":{"userId":"User id"}}}}}},"responses":{"200":{"description":"lesson join response","content":{"application/json":{"schema":{"type":"string"}}}},"400":{"description":"Parameters are not valid"}},"tags":["Lesson"]}},"/question":{"post":{"operationId":"QuestionController_registerQuestion","parameters":[],"requestBody":{"required":true,"description":"user data model","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateQuestionModel"},"examples":{"template":{"value":{"lessonId":"Lesson Id","title":"Question Title","description":"Question decsription","data":"Question data","typeAnswer":"TEXT","typeQuestion":"TEXT","picture":"Question picture","difficulty":"BEGINNER"}}}}}},"responses":{"200":{"description":"question create response","content":{"application/json":{"schema":{"type":"string"}}}},"400":{"description":"Parameters are not valid"}},"tags":["Question"]},"delete":{"operationId":"QuestionController_deleteQuestion","parameters":[],"requestBody":{"required":true,"description":"user data model","content":{"application/json":{"schema":{"$ref":"#/components/schemas/IdQuestionModel"},"examples":{"template":{"value":{"id":"id"}}}}}},"responses":{"200":{"description":"question delete response","content":{"application/json":{"schema":{"type":"string"}}}},"400":{"description":"Parameters are not valid"}},"tags":["Question"]}},"/question/{id}":{"get":{"operationId":"QuestionController_getQuestion","parameters":[{"name":"id","required":true,"in":"path","description":"Id of the question","schema":{"type":"string"}}],"responses":{"200":{"description":"question content response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/QuestionModel"}}}},"400":{"description":"Parameters are not valid"}},"tags":["Question"]},"put":{"operationId":"QuestionController_updateQuestion","parameters":[{"name":"id","required":true,"in":"path","description":"Id of the question","schema":{"type":"string"}}],"requestBody":{"required":true,"description":"user data model","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateQuestionModel"},"examples":{"template":{"value":{"lessonId":"id","title":"Question Title","description":"Question decsription","data":"Data of the question","trustAnswerId":"id"}}}}}},"responses":{"200":{"description":"question update response","content":{"application/json":{"schema":{"type":"string"}}}},"400":{"description":"Parameters are not valid"}},"tags":["Question"]}},"/question/{id}/answers":{"get":{"operationId":"QuestionController_getQuestionAnswers","parameters":[{"name":"id","required":true,"in":"path","description":"Id of the question","schema":{"type":"string"}}],"responses":{"200":{"description":"question content response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/QuestionModel"}}}},"400":{"description":"Parameters are not valid"}},"tags":["Question"]}},"/answer":{"post":{"operationId":"AnswerController_registerAnswer","parameters":[],"requestBody":{"required":true,"description":"user data model","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateAnswerModel"},"examples":{"template":{"value":{"questionId":"Question Id","data":"Answer data","picture":"Picture data"}}}}}},"responses":{"200":{"description":"answer create response","content":{"application/json":{"schema":{"type":"string"}}}},"400":{"description":"Parameters are not valid"}},"tags":["Answer"]},"delete":{"operationId":"AnswerController_deleteAnswer","parameters":[],"requestBody":{"required":true,"description":"user data model","content":{"application/json":{"schema":{"$ref":"#/components/schemas/IdAnswerModel"},"examples":{"template":{"value":{"id":"id"}}}}}},"responses":{"200":{"description":"answer delete response","content":{"application/json":{"schema":{"type":"string"}}}},"400":{"description":"Parameters are not valid"}},"tags":["Answer"]}},"/answer/{id}":{"get":{"operationId":"AnswerController_getAnswer","parameters":[{"name":"id","required":true,"in":"path","description":"Id of the answer","schema":{"type":"string"}}],"responses":{"200":{"description":"answer content response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AnswerModel"}}}},"400":{"description":"Parameters are not valid"}},"tags":["Answer"]},"put":{"operationId":"AnswerController_updateAnswer","parameters":[{"name":"id","required":true,"in":"path","description":"Id of the answer","schema":{"type":"string"}}],"requestBody":{"required":true,"description":"user data model","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateAnswerModel"},"examples":{"template":{"value":{"questionId":"Question Id","data":"Answer data","picture":"Picture data"}}}}}},"responses":{"200":{"description":"answer update response","content":{"application/json":{"schema":{"type":"string"}}}},"400":{"description":"Parameters are not valid"}},"tags":["Answer"]}}},"info":{"title":"Ollamy API","description":"So insane API","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"cookie":{"type":"apiKey","in":"cookie","name":"session"}},"schemas":{"CreateUserModel":{"type":"object","properties":{"firstname":{"type":"string"},"lastname":{"type":"string"},"email":{"type":"string"},"password":{"type":"string","description":"Password must contain at least 8 characters, 2 numbers and 2 uppercase letters"}},"required":["firstname","lastname","email","password"]},"LoginUserModel":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}},"required":["email","password"]},"GetUserModel":{"type":"object","properties":{"firstname":{"type":"string"},"lastname":{"type":"string"},"email":{"type":"string"}},"required":["firstname","lastname","email"]},"UpdateUserModel":{"type":"object","properties":{"firstname":{"type":"string"},"lastname":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"}},"required":["firstname","lastname","email","password"]},"CreateCourseModel":{"type":"object","properties":{"title":{"type":"string"},"description":{"type":"string"},"picture":{"type":"string"}},"required":["title","description","picture"]},"IdCourseModel":{"type":"object","properties":{"id":{"type":"string"}},"required":["id"]},"CourseModel":{"type":"object","properties":{"id":{"type":"string"},"ownerId":{"type":"string"},"title":{"type":"string"},"description":{"type":"string"},"picture":{"type":"string"}},"required":["id","ownerId","title","description","picture"]},"UpdateCourseModel":{"type":"object","properties":{"ownerId":{"type":"string"},"title":{"type":"string"},"description":{"type":"string"},"picture":{"type":"string"}},"required":["ownerId","title","description","picture"]},"SectionModel":{"type":"object","properties":{"id":{"type":"string"},"courseId":{"type":"string"},"title":{"type":"string"},"description":{"type":"string"}},"required":["id","courseId","title","description"]},"CreateSectionModel":{"type":"object","properties":{"courseId":{"type":"string"},"title":{"type":"string"},"description":{"type":"string"}},"required":["courseId","title","description"]},"IdSectionModel":{"type":"object","properties":{"id":{"type":"string"}},"required":["id"]},"UpdateSectionModel":{"type":"object","properties":{"courseId":{"type":"string"},"title":{"type":"string"},"description":{"type":"string"}},"required":["courseId","title","description"]},"LessonModel":{"type":"object","properties":{"id":{"type":"string"},"sectionId":{"type":"string"},"title":{"type":"string"},"description":{"type":"string"}},"required":["id","sectionId","title","description"]},"CreateLessonModel":{"type":"object","properties":{"section_id":{"type":"string"},"title":{"type":"string"},"description":{"type":"string"}},"required":["section_id","title","description"]},"IdLessonModel":{"type":"object","properties":{"id":{"type":"string"}},"required":["id"]},"QuestionModel":{"type":"object","properties":{"id":{"type":"string"},"lessonId":{"type":"string"},"title":{"type":"string"},"description":{"type":"string"},"typeAnswer":{"type":"string"},"typeQuestion":{"type":"string"},"trustAnswerId":{"type":"string"},"pictureId":{"type":"string"},"difficulty":{"type":"string"}},"required":["id","lessonId","title","description","typeAnswer","typeQuestion","trustAnswerId","pictureId","difficulty"]},"JoinLessonModel":{"type":"object","properties":{"userId":{"type":"string"}},"required":["userId"]},"CreateQuestionModel":{"type":"object","properties":{"lessonId":{"type":"string"},"title":{"type":"string"},"description":{"type":"string"},"data":{"type":"string"},"typeAnswer":{"type":"string"},"typeQuestion":{"type":"string"},"picture":{"type":"string"},"difficulty":{"type":"string"}},"required":["lessonId","title","description","data","typeAnswer","typeQuestion","picture","difficulty"]},"IdQuestionModel":{"type":"object","properties":{"id":{"type":"string"}},"required":["id"]},"UpdateQuestionModel":{"type":"object","properties":{"lessonId":{"type":"string"},"title":{"type":"string"},"description":{"type":"string"},"picture":{"type":"string"},"points":{"type":"number"},"difficulty":{"type":"string"},"trustAnswerId":{"type":"string"}},"required":["lessonId","title","description","picture","points","difficulty","trustAnswerId"]},"CreateAnswerModel":{"type":"object","properties":{"questionId":{"type":"string"},"data":{"type":"string"},"picture":{"type":"string"}},"required":["questionId","data","picture"]},"IdAnswerModel":{"type":"object","properties":{"id":{"type":"string"}},"required":["id"]},"AnswerModel":{"type":"object","properties":{"id":{"type":"string"},"questionId":{"type":"string"},"data":{"type":"string"},"picture":{"type":"string"}},"required":["id","questionId","data","picture"]},"UpdateAnswerModel":{"type":"object","properties":{"questionId":{"type":"string"},"data":{"type":"string"},"picture":{"type":"string"}},"required":["questionId","data","picture"]}}}}